update to handle custom structs as property
it doesn't look like unium handles custom structs as a property to a
component when retrieving the data for the component. this means they
need to be retrieved separately. this update should also handle at a
basic level component properties that is a custom struct,
which itself has custom structs as properties.