remove expand/collapse option for props
this was done for custom struct props which required separate calls to
get its data. it was put in because of the potential for cyclical
references which would have been a problem if the child props were
automatically retrieved. however, some "set" queries require all data to
be submitted which means if you don't expand a sibling property when
submitting an update for another subling, the former sibling will be
cleared, because no data was submitted for it. so bear in mind that the
cyclical reference issue could be a thing.